// src/shared/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  logistics: {
    Tables: {
// ...
      items: {
        Row: {
          base_uom: string
          category: string | null
          created_at: string
          created_by: string | null
          currency: string | null
          description: string | null
          dimensions_json: Json | null
          expiry_tracked: boolean
          id: string
          lot_tracked: boolean
          name: string
          serial_tracked: boolean
          sku_code: string
          standard_cost: number | null
          status: string
          tenant_id: string
          type: string
          updated_at: string
          updated_by: string | null
          weight_kg: number | null
        }
        Insert: {
          base_uom?: string
          category?: string | null
          created_at?: string
          created_by?: string | null
          currency?: string | null
          description?: string | null
          dimensions_json?: Json | null
          expiry_tracked?: boolean
          id?: string
          lot_tracked?: boolean
          name: string
          serial_tracked?: boolean
          sku_code: string
          standard_cost?: number | null
          status?: string
          tenant_id: string
          type?: string
          updated_at?: string
          updated_by?: string | null
          weight_kg?: number | null
        }
        Update: {
          base_uom?: string
          category?: string | null
          created_at?: string
          created_by?: string | null
          currency?: string | null
          description?: string | null
          dimensions_json?: Json | null
          expiry_tracked?: boolean
          id?: string
          lot_tracked?: boolean
          name?: string
          serial_tracked?: boolean
          sku_code?: string
          standard_cost?: number | null
          status?: string
          tenant_id?: string
          type?: string
          updated_at?: string
          updated_by?: string | null
          weight_kg?: number | null
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}